requirements.txt
python-docx
spacy
transformers
torch
openai
fastapi
uvicorn
pdfplumber
streamlit
python-dotenv
python-multipart
pydantic
requests
